+++ bin/tasks/post-block.sh
@@ -2,14 +2,20 @@
 # @Author: Roni Laukkarinen
 # @Date:   2022-02-08 16:03:52
 # @Last Modified by:   Roni Laukkarinen
-# @Last Modified time: 2022-02-22 12:35:21
+# @Last Modified time: 2022-02-24 16:44:27
 # Tasks after functions.php has been updated
+
+# Let's go to the project directory
+cd ${PROJECTS_HOME}/${PROJECT_NAME}
 
 # Remove the file without any changes
 rm ${PROJECT_THEME_PATH}/functions.php
 
 # Rename the changed file to the official one
 mv ${PROJECT_THEME_PATH}/tmpfile ${PROJECT_THEME_PATH}/functions.php
+
+# Ensure that tmpfiles are removed
+rm ${PROJECT_THEME_PATH}/tmpfile
 
 # Run required things after the block/<block>.sh
 # (like localization.sh that needs to be run after the <block>.sh
@@ -22,13 +28,6 @@ echo "${BOLDYELLOW}Importing ACF fields (json)...${TXTRESET} "
 if [[ ${AIR_BLOCKS_LANG} = "en" ]]; then
   # Run localization task
   source ${SCRIPTS_LOCATION}/tasks/localization.sh
-
-  # Tasks after functions.php has been updated, again (in localization.sh)
-  # Remove the file without any changes
-  rm ${PROJECT_THEME_PATH}/functions.php
-
-  # Rename the changed file to the official one
-  mv ${PROJECT_THEME_PATH}/tmpfile ${PROJECT_THEME_PATH}/functions.php
 fi
 
 # Just import the ACF fields file